+++ lib/vagrant/message.ex
@@ -1,15 +1,14 @@
 defmodule Vagrant.Message do
-
   alias Vagrant.Message.MachineReadable
-  alias Vagrant.Message.Status
-
+  
   def parse(cmd, {:ok, result}) do
     result =
       try do
         mod = Module.concat(Vagrant.Message, String.capitalize(cmd))
         apply(mod, :load, [result])
       rescue
-        _e -> result
+        _e ->
+          String.strip(result)
       end
     {:ok, result}
   end