+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 # v0.0.2-dev
 ## Enhancements
 * Added version / installed? methods for safely querying for the availability of vagrant
+* Added ability to perform ssh commands on the vagrant session
 
 # v0.0.1
 init